// apps/tenant-admin/api/src/modules/delivery/contracts/delivery-errors.ts
export enum DeliveryErrorCode {
    FeatureDisabled = 'DELIVERY_FEATURE_DISABLED',
    NotFound = 'DELIVERY_NOT_FOUND',
    TenantMismatch = 'DELIVERY_TENANT_MISMATCH',
    BatchNotDelivery = 'DELIVERY_BATCH_NOT_DELIVERY',
    AlreadyExists = 'DELIVERY_ALREADY_EXISTS',
    InvalidStatusTransition = 'DELIVERY_INVALID_STATUS_TRANSITION',
    VersionConflict = 'DELIVERY_VERSION_CONFLICT',
    IdempotencyConflict = 'DELIVERY_IDEMPOTENCY_CONFLICT',
    InvalidCommand = 'DELIVERY_INVALID_COMMAND',
    AddressNotConfirmed = 'DELIVERY_ADDRESS_NOT_CONFIRMED',
    AddressOutsideArea = 'DELIVERY_ADDRESS_OUTSIDE_AREA',
    GeocodeUnavailable = 'DELIVERY_GEOCODE_UNAVAILABLE',
    DriverNotFound = 'DELIVERY_DRIVER_NOT_FOUND',
    DriverNotAvailable = 'DELIVERY_DRIVER_NOT_AVAILABLE',
    DriverBusy = 'DELIVERY_DRIVER_BUSY',
    AssignmentNotAllowed = 'DELIVERY_ASSIGNMENT_NOT_ALLOWED',
    NotAssignedDriver = 'DELIVERY_NOT_ASSIGNED_DRIVER',
    LocationInvalid = 'DELIVERY_LOCATION_INVALID',
    LocationStale = 'DELIVERY_LOCATION_STALE',
    LocationDuplicate = 'DELIVERY_LOCATION_DUPLICATE',
    TrackingAccessInvalid = 'DELIVERY_TRACKING_ACCESS_INVALID',
    TrackingAccessExpired = 'DELIVERY_TRACKING_ACCESS_EXPIRED',
    PinRequired = 'DELIVERY_PIN_REQUIRED',
    PinInvalid = 'DELIVERY_PIN_INVALID',
    PinLocked = 'DELIVERY_PIN_LOCKED',
    OverrideNotAllowed = 'DELIVERY_OVERRIDE_NOT_ALLOWED',
    OverrideReasonRequired = 'DELIVERY_OVERRIDE_REASON_REQUIRED',
    ExceptionInvalid = 'DELIVERY_EXCEPTION_INVALID',
    RateLimited = 'DELIVERY_RATE_LIMITED',
    CustomerNotFound = 'DELIVERY_CUSTOMER_NOT_FOUND',
    CustomerPhoneInvalid = 'DELIVERY_CUSTOMER_PHONE_INVALID',
    AddressLimitExceeded = 'DELIVERY_ADDRESS_LIMIT_EXCEEDED',
    AddressNotFound = 'DELIVERY_ADDRESS_NOT_FOUND',
    AddressDeleted = 'DELIVERY_ADDRESS_DELETED',
    PostalCodeInvalid = 'DELIVERY_POSTAL_CODE_INVALID',
    PostalCodeNotFound = 'DELIVERY_POSTAL_CODE_NOT_FOUND',
    PostalCodeUnavailable = 'DELIVERY_POSTAL_CODE_UNAVAILABLE',
    GeocodeAmbiguous = 'DELIVERY_GEOCODE_AMBIGUOUS',
    CapacityUnavailable = 'DELIVERY_OWN_CAPACITY_UNAVAILABLE',
    CapacityReservationExpired = 'DELIVERY_CAPACITY_RESERVATION_EXPIRED',
    CapacityReservationConflict = 'DELIVERY_CAPACITY_RESERVATION_CONFLICT',
    QuoteNotFound = 'DELIVERY_QUOTE_NOT_FOUND',
    QuoteExpired = 'DELIVERY_QUOTE_EXPIRED',
    QuoteAlreadyUsed = 'DELIVERY_QUOTE_ALREADY_USED',
    FulfillmentNotFound = 'DELIVERY_FULFILLMENT_NOT_FOUND',
    ProviderNotConfigured = 'DELIVERY_PROVIDER_NOT_CONFIGURED',
    ProviderCredentialInvalid = 'DELIVERY_PROVIDER_CREDENTIAL_INVALID',
    ProviderUnavailable = 'DELIVERY_PROVIDER_UNAVAILABLE',
    ProviderCycleExhausted = 'DELIVERY_PROVIDER_CYCLE_EXHAUSTED',
    ProviderSwitchNotAllowed = 'DELIVERY_PROVIDER_SWITCH_NOT_ALLOWED',
    FulfillmentAlreadyCollected = 'DELIVERY_FULFILLMENT_ALREADY_COLLECTED',
    FinancialSnapshotConflict = 'DELIVERY_FINANCIAL_SNAPSHOT_CONFLICT',
}

/**
 * Default HTTP mapping. Domain services may override a status when returning
 * more context, but the code itself must remain stable for clients.
 */
export const DELIVERY_ERROR_HTTP_STATUS: Readonly<Record<DeliveryErrorCode, HttpStatus>> = {
    [DeliveryErrorCode.FeatureDisabled]: HttpStatus.NOT_FOUND,
    [DeliveryErrorCode.NotFound]: HttpStatus.NOT_FOUND,
    [DeliveryErrorCode.TenantMismatch]: HttpStatus.FORBIDDEN,
    [DeliveryErrorCode.BatchNotDelivery]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.AlreadyExists]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.InvalidStatusTransition]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.VersionConflict]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.IdempotencyConflict]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.InvalidCommand]: HttpStatus.BAD_REQUEST,
    [DeliveryErrorCode.AddressNotConfirmed]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.AddressOutsideArea]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.GeocodeUnavailable]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.DriverNotFound]: HttpStatus.NOT_FOUND,
    [DeliveryErrorCode.DriverNotAvailable]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.DriverBusy]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.AssignmentNotAllowed]: HttpStatus.FORBIDDEN,
    [DeliveryErrorCode.NotAssignedDriver]: HttpStatus.FORBIDDEN,
    [DeliveryErrorCode.LocationInvalid]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.LocationStale]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.LocationDuplicate]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.TrackingAccessInvalid]: HttpStatus.UNAUTHORIZED,
    [DeliveryErrorCode.TrackingAccessExpired]: HttpStatus.UNAUTHORIZED,
    [DeliveryErrorCode.PinRequired]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.PinInvalid]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.PinLocked]: HttpStatus.TOO_MANY_REQUESTS,
    [DeliveryErrorCode.OverrideNotAllowed]: HttpStatus.FORBIDDEN,
    [DeliveryErrorCode.OverrideReasonRequired]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.ExceptionInvalid]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.RateLimited]: HttpStatus.TOO_MANY_REQUESTS,
    [DeliveryErrorCode.CustomerNotFound]: HttpStatus.NOT_FOUND,
    [DeliveryErrorCode.CustomerPhoneInvalid]: HttpStatus.BAD_REQUEST,
    [DeliveryErrorCode.AddressLimitExceeded]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.AddressNotFound]: HttpStatus.NOT_FOUND,
    [DeliveryErrorCode.AddressDeleted]: HttpStatus.GONE,
    [DeliveryErrorCode.PostalCodeInvalid]: HttpStatus.BAD_REQUEST,
    [DeliveryErrorCode.PostalCodeNotFound]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.PostalCodeUnavailable]: HttpStatus.BAD_GATEWAY,
    [DeliveryErrorCode.GeocodeAmbiguous]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.CapacityUnavailable]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.CapacityReservationExpired]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.CapacityReservationConflict]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.QuoteNotFound]: HttpStatus.NOT_FOUND,
    [DeliveryErrorCode.QuoteExpired]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.QuoteAlreadyUsed]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.FulfillmentNotFound]: HttpStatus.NOT_FOUND,
    [DeliveryErrorCode.ProviderNotConfigured]: HttpStatus.UNPROCESSABLE_ENTITY,
    [DeliveryErrorCode.ProviderCredentialInvalid]: HttpStatus.UNAUTHORIZED,
    [DeliveryErrorCode.ProviderUnavailable]: HttpStatus.BAD_GATEWAY,
    [DeliveryErrorCode.ProviderCycleExhausted]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.ProviderSwitchNotAllowed]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.FulfillmentAlreadyCollected]: HttpStatus.CONFLICT,
    [DeliveryErrorCode.FinancialSnapshotConflict]: HttpStatus.CONFLICT,
};